Fixes for RequestStatus throwing Rack::Lint::LintError: errors
'No Content-Type header found' was being thrown due to an empty {}
being returned.

'Response body must respond to each' was being thrown due to "" no
longer responding to each as it did in ruby 1.8.7.

Both of these fixes are necessary for the call(env) method on the RequestStatus
class to respond properly. The gem currently does not have a test suite
to verify these changes, but I've confirmed each example in the comments for
RequestStatus responds correctly.